Saving works for New entry and New Billings

+++ src/main/java/com/khrizman/pledger/repository/LedgerEntryRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface LedgerEntryRepository extends JpaRepository<LedgerEntry, Integer> {
+public interface LedgerEntryRepository extends JpaRepository<LedgerEntry, Long> {
     List<LedgerEntry> findAllByEntryCompleteOrderByStartDate(boolean complete);
 
     LedgerEntry findById(int id);
